refactored custom fabric with browserify
