+++ public/js/fabric_custom/elements/LabeledCanvas.js
@@ -1,0 +1,43 @@
+// impt 
+var fabric = require('fabric').fabric;
+
+// TODO: BAD make this a var declaration. Global for now because welcome controller depends on fabric.LabeledCanvas
+LabeledCanvas = fabric.util.createClass(fabric.Canvas, {
+  initialize: function(el, options) {
+    options || (options = { });
+    this.callSuper('initialize', el, options);
+    this._id = options._id;
+    this.name = options.name;
+    this.type = 'labeled-canvas';
+  },
+  toObject: function() {
+    return fabric.util.object.extend(this.callSuper('toObject'), {
+      _id: this._id,
+      name: this.name,
+      type: 'labeled-canvas'
+    });
+  },
+  toJSON: function() {
+    return fabric.util.object.extend(this.callSuper('toJSON'), {
+      _id: this._id,
+      name: this.name,
+      type: 'labeled-canvas'
+    });
+  }
+});
+
+/**
+ *
+ * http://jsfiddle.net/rwaldron/j3vST/
+ *
+ * @param id
+ * @returns {*}
+ */
+LabeledCanvas.prototype.findById = function(id){
+  return this.getObjects().filter(function( obj ) {
+    return obj._id === id;
+  })[ 0 ];
+}
+
+module.exports = LabeledCanvas;
+
