+++ public/js/fabric_custom/elements/LabeledRect.js
@@ -1,0 +1,22 @@
+var fabric = require('fabric').fabric;
+var LabeledRect = fabric.util.createClass(fabric.Rect, {
+  type: 'labeled-rect',
+  initialize: function(options) {
+    options || (options = { });
+    this.callSuper('initialize', options);
+    this.set('_id', options._id);
+    this.set('drawingId', options.drawingId);
+  },
+  toObject: function() {
+    return fabric.util.object.extend(this.callSuper('toObject'), {
+      _id: this.get('_id'),
+      drawingId: this.get('drawingId')
+    });
+  }
+});
+
+LabeledRect.fromObject = function(o, callback){
+  return new LabeledRect(o);
+};
+
+module.exports = LabeledRect;
