+++ public/js/fabric_custom/elements/LabeledTriangle.js
@@ -1,0 +1,22 @@
+var fabric = require('fabric').fabric;
+var LabeledTriangle = fabric.util.createClass(fabric.Triangle, {
+  type: 'labeled-triangle',
+  initialize: function(options) {
+    options || (options = { });
+    this.callSuper('initialize', options);
+    this.set('_id', options._id);
+    this.set('drawingId', options.drawingId);
+  },
+  toObject: function() {
+    return fabric.util.object.extend(this.callSuper('toObject'), {
+      _id: this.get('_id'),
+      drawingId: this.get('drawingId')
+    });
+  }
+});
+
+LabeledTriangle.fromObject = function(o, callback){
+  return new LabeledTriangle(o);
+};
+
+module.exports = LabeledTriangle;
