+++ public/js/services/sessions.js
@@ -1,6 +1,5 @@
 'use strict';
 
-angular.module('users')
-  .factory('Session', function ($resource) {
-    return $resource('/auth/session/');
-  });
+angular.module('users').factory('Session', function ($resource) {
+  return $resource('/auth/session/');
+});
